@@ -1,16 +1,15 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
-namespace DotNet.GitHubAction.Extensions
-{
-    static class ServiceCollectionExtensions
-    {
-        internal static IServiceCollection AddGitHubActionServices(
-            this IServiceCollection services)
-        {
-            // Omitted for brevity
-            // https://github.com/dotnet/samples/blob/main/github-actions/DotNet.GitHubAction/DotNet.GitHubAction/Extensions/ServiceCollectionExtensions.cs
+namespace DotNet.GitHubAction.Extensions;
 
-            return services;
-        }
+static class ServiceCollectionExtensions
+{
+    internal static IServiceCollection AddGitHubActionServices(
+        this IServiceCollection services)
+    {
+        // Omitted for brevity
+        // https://github.com/dotnet/samples/blob/main/github-actions/DotNet.GitHubAction/DotNet.GitHubAction/Extensions/ServiceCollectionExtensions.cs
+
+        return services;
     }
 }
