@@ -1,4 +1,5 @@
-﻿using BasicHttp.Example;
+﻿using Shared;
+using BasicHttp.Example;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -7,18 +8,24 @@ using IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureServices(services =>
     {
         services.AddHttpClient();
-        services.AddTransient<JokeService>();
+        services.AddTransient<TodoService>();
     })
     .Build();
 
-JokeService jokeService =
-    host.Services.GetRequiredService<JokeService>();
+TodoService todoService =
+    host.Services.GetRequiredService<TodoService>();
 
-string jokeText = await jokeService.GetRandomJokeAsync();
+Todo[] todos = await todoService.GetUserTodosAsync(1);
 
 ILogger logger =
-    host.Services.GetRequiredService<ILogger<JokeService>>();
+    host.Services.GetRequiredService<ILogger<TodoService>>();
 
-logger.LogInformation("Joke: {Text}", jokeText);
+foreach (Todo? todo in todos)
+{
+    logger.LogInformation("Todo: {Details}", $"""
+        Id: {todo?.Id} (Is completed: {todo?.Completed})
+        Title: {todo?.Title}
+        """);
+}
 
 await host.RunAsync();
