@@ -1,4 +1,5 @@
-﻿using NamedHttp.Example;
+﻿using Shared;
+using NamedHttp.Example;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -6,29 +7,35 @@ using Microsoft.Extensions.Logging;
 using IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureServices((context, services) =>
     {
-        string? httpClientName = context.Configuration["JokeHttpClientName"];
+        string? httpClientName = context.Configuration["TodoHttpClientName"];
         services.AddHttpClient(
             httpClientName ?? "",
             client =>
             {
                 // Set the base address of the named client.
-                client.BaseAddress = new Uri("https://api.icndb.com/");
+                client.BaseAddress = new Uri("https://jsonplaceholder.typicode.com/");
 
                 // Add a user-agent default request header.
                 client.DefaultRequestHeaders.UserAgent.ParseAdd("dotnet-docs");
             });
-        services.AddTransient<JokeService>();
+        services.AddTransient<TodoService>();
     })
     .Build();
 
-JokeService jokeService =
-    host.Services.GetRequiredService<JokeService>();
+TodoService todoService =
+    host.Services.GetRequiredService<TodoService>();
 
-string jokeText = await jokeService.GetRandomJokeAsync();
+Todo[] todos = await todoService.GetUserTodosAsync(3);
 
 ILogger logger =
-    host.Services.GetRequiredService<ILogger<JokeService>>();
+    host.Services.GetRequiredService<ILogger<TodoService>>();
 
-logger.LogInformation("Joke: {Text}", jokeText);
+foreach (Todo? todo in todos)
+{
+    logger.LogInformation("Todo: {Details}", $"""
+        Id: {todo?.Id} (Is completed: {todo?.Completed})
+        Title: {todo?.Title}
+        """);
+}
 
 await host.RunAsync();
