@@ -1,4 +1,5 @@
-﻿using TypedHttp.Example;
+﻿using Shared;
+using TypedHttp.Example;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -6,11 +7,11 @@ using Microsoft.Extensions.Logging;
 using IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureServices(services =>
     {
-        services.AddHttpClient<JokeService>(
+        services.AddHttpClient<TodoService>(
             client =>
             {
                 // Set the base address of the named client.
-                client.BaseAddress = new Uri("https://api.icndb.com/");
+                client.BaseAddress = new Uri("https://jsonplaceholder.typicode.com/");
 
                 // Add a user-agent default request header.
                 client.DefaultRequestHeaders.UserAgent.ParseAdd("dotnet-docs");
@@ -18,14 +19,20 @@ using IHost host = Host.CreateDefaultBuilder(args)
     })
     .Build();
 
-JokeService jokeService =
-    host.Services.GetRequiredService<JokeService>();
+TodoService todoService =
+    host.Services.GetRequiredService<TodoService>();
 
-string jokeText = await jokeService.GetRandomJokeAsync();
+Todo[] todos = await todoService.GetUserTodosAsync(4);
 
 ILogger logger =
-    host.Services.GetRequiredService<ILogger<JokeService>>();
+    host.Services.GetRequiredService<ILogger<TodoService>>();
 
-logger.LogInformation("Joke: {Text}", jokeText);
+foreach (Todo? todo in todos)
+{
+    logger.LogInformation("Todo: {Details}", $"""
+        Id: {todo?.Id} (Is completed: {todo?.Completed})
+        Title: {todo?.Title}
+        """);
+}
 
 await host.RunAsync();
