@@ -42,7 +42,7 @@ namespace StringLibraryTest
             string?[] words = { string.Empty, null };
             foreach (var word in words)
             {
-                bool result = StringLibrary.StartsWithUpper(word ?? string.Empty);
+                bool result = StringLibrary.StartsWithUpper(word);
                 Assert.IsFalse(result,
                        string.Format("Expected for '{0}': false; Actual: {1}",
                                      word == null ? "<null>" : word, result));
