@@ -1,6 +1,6 @@
 ﻿using System;
 
-public enum CoffeChoice
+public enum CoffeeChoice
 {
     Plain,
     WithMilk,
@@ -11,27 +11,27 @@ public class GotoInSwitchExample
 {
     public static void Main()
     {
-        Console.WriteLine(CalculatePrice(CoffeChoice.Plain));  // output: 10.0
-        Console.WriteLine(CalculatePrice(CoffeChoice.WithMilk));  // output: 15.0
-        Console.WriteLine(CalculatePrice(CoffeChoice.WithIceCream));  // output: 17.0
+        Console.WriteLine(CalculatePrice(CoffeeChoice.Plain));  // output: 10.0
+        Console.WriteLine(CalculatePrice(CoffeeChoice.WithMilk));  // output: 15.0
+        Console.WriteLine(CalculatePrice(CoffeeChoice.WithIceCream));  // output: 17.0
     }
 
-    private static decimal CalculatePrice(CoffeChoice choice)
+    private static decimal CalculatePrice(CoffeeChoice choice)
     {
         decimal price = 0;
         switch (choice)
         {
-            case CoffeChoice.Plain:
+            case CoffeeChoice.Plain:
                 price += 10.0m;
                 break;
 
-            case CoffeChoice.WithMilk:
+            case CoffeeChoice.WithMilk:
                 price += 5.0m;
-                goto case CoffeChoice.Plain;
+                goto case CoffeeChoice.Plain;
 
-            case CoffeChoice.WithIceCream:
+            case CoffeeChoice.WithIceCream:
                 price += 7.0m;
-                goto case CoffeChoice.Plain;
+                goto case CoffeeChoice.Plain;
         }
         return price;
     }
