@@ -1,52 +1,44 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Text;
+﻿using System.ComponentModel;
+using System.Text.Json.Serialization;
 
 namespace serialization
 {
-    // <Snippet2>
-    [Serializable()]
-    // </Snippet2>
     // <Snippet1>
-   public class Loan : INotifyPropertyChanged
+    public class Loan : INotifyPropertyChanged
     {
         public double LoanAmount { get; set; }
-        public double InterestRatePercent { get; set; }
+        public double InterestRate { get; set; }
 
-        // <Snippet4>
-        [field:NonSerialized()]
+        // <Snippet2>
+        [JsonIgnore]
         public DateTime TimeLastLoaded { get; set; }
-        // </Snippet4>
+        // </Snippet2>
 
         public int Term { get; set; }
 
-        private string customer;
+        private string _customer;
         public string Customer
         {
-            get { return customer; }
+            get { return _customer; }
             set
             {
-                customer = value;
+                _customer = value;
                 PropertyChanged?.Invoke(this,
                   new PropertyChangedEventArgs(nameof(Customer)));
             }
         }
 
-        // <Snippet3>
-        [field: NonSerialized()]
-        public event System.ComponentModel.PropertyChangedEventHandler PropertyChanged;
-        // </Snippet3>
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         public Loan(double loanAmount,
                     double interestRate,
                     int term,
                     string customer)
         {
-            this.LoanAmount = loanAmount;
-            this.InterestRatePercent = interestRate;
-            this.Term = term;
-            this.customer = customer;
+            LoanAmount = loanAmount;
+            InterestRate = interestRate;
+            Term = term;
+            _customer = customer;
         }
     }
     // </Snippet1>
