@@ -1,7 +1,7 @@
 ﻿'------------------------------------------------------------------------------
 ' <auto-generated>
 '     This code was generated by a tool.
-'     Runtime Version:4.0.20926.1
+'     Runtime Version:4.0.30319.42000
 '
 '     Changes to this file may cause incorrect behavior and will be lost if
 '     the code is regenerated.
