@@ -1,35 +1,36 @@
 ﻿'------------------------------------------------------------------------------
-' <autogenerated>
+' <auto-generated>
 '     This code was generated by a tool.
-'     Runtime Version:2.0.50107.0
+'     Runtime Version:4.0.30319.42000
 '
 '     Changes to this file may cause incorrect behavior and will be lost if
 '     the code is regenerated.
-' </autogenerated>
+' </auto-generated>
 '------------------------------------------------------------------------------
 
 Option Strict On
 Option Explicit On
 
+
 Namespace My
     
-    'NOTE: This file is auto-generated, do not modify it directly.  To make changes,
+    'NOTE: This file is auto-generated; do not modify it directly.  To make changes,
     ' or if you encounter build errors in this file, go to the Project Designer
-    ' (go to project properties or double-click on the My Project node in the 
+    ' (go to Project Properties or double-click the My Project node in
     ' Solution Explorer), and make changes on the Application tab.
     '
-    Partial Class MyApplication
+    Partial Friend Class MyApplication
         
-        <Global.System.Diagnostics.DebuggerStepThrough()>  _
+        <Global.System.Diagnostics.DebuggerStepThroughAttribute()>  _
         Public Sub New()
-            MyBase.New(Microsoft.VisualBasic.ApplicationServices.AuthenticationMode.Windows)
+            MyBase.New(Global.Microsoft.VisualBasic.ApplicationServices.AuthenticationMode.Windows)
             Me.IsSingleInstance = false
             Me.EnableVisualStyles = true
             Me.SaveMySettingsOnExit = true
-            Me.ShutDownStyle = Microsoft.VisualBasic.ApplicationServices.ShutdownMode.AfterMainFormCloses
+            Me.ShutDownStyle = Global.Microsoft.VisualBasic.ApplicationServices.ShutdownMode.AfterMainFormCloses
         End Sub
         
-        <Global.System.Diagnostics.DebuggerStepThrough()>  _
+        <Global.System.Diagnostics.DebuggerStepThroughAttribute()>  _
         Protected Overrides Sub OnCreateMainForm()
             Me.MainForm = Global.VbVbalrOOP.Form1
         End Sub
