@@ -1,65 +1,72 @@
 ﻿'------------------------------------------------------------------------------
-' <autogenerated>
+' <auto-generated>
 '     This code was generated by a tool.
-'     Runtime Version:2.0.50107.0
+'     Runtime Version:4.0.30319.42000
 '
 '     Changes to this file may cause incorrect behavior and will be lost if
 '     the code is regenerated.
-' </autogenerated>
+' </auto-generated>
 '------------------------------------------------------------------------------
 
 Option Strict On
 Option Explicit On
 
-<Assembly: Global.System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode", Scope:="member", Target:="VbVbalrOOP.Settings.get_Default():VbVbalrOOP.Settings"),  _
- Assembly: Global.System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode", Scope:="member", Target:="VbVbalrOOP.My.MySettingsProperty.get_Settings():VbVbalrOOP.Settings")> 
 
-
-<Global.System.ComponentModel.EditorBrowsableAttribute(Global.System.ComponentModel.EditorBrowsableState.Advanced)>  _
-Partial NotInheritable Class Settings
-    Inherits Global.System.Configuration.ApplicationSettingsBase
+Namespace My
     
-    <Global.System.Runtime.CompilerServices.CompilerGeneratedAttribute()>  _
-    Private Shared defaultInstance As Settings = New Settings
-    
+    <Global.System.Runtime.CompilerServices.CompilerGeneratedAttribute(),  _
+     Global.System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.4.0.0"),  _
+     Global.System.ComponentModel.EditorBrowsableAttribute(Global.System.ComponentModel.EditorBrowsableState.Advanced)>  _
+    Partial Friend NotInheritable Class Settings
+        Inherits Global.System.Configuration.ApplicationSettingsBase
+        
+        Private Shared defaultInstance As Settings = CType(Global.System.Configuration.ApplicationSettingsBase.Synchronized(New Settings()),Settings)
+        
 #Region "My.Settings Auto-Save Functionality"
 #If _MyType = "WindowsForms" Then
-    <Global.System.Runtime.CompilerServices.CompilerGeneratedAttribute()>  _
     Private Shared addedHandler As Boolean
 
-    <Global.System.Diagnostics.DebuggerNonUserCodeAttribute(), Global.System.ComponentModel.EditorBrowsableAttribute(Global.System.ComponentModel.EditorBrowsableState.Advanced)> _ 
-    Private Shared Sub AutoSaveSettings(ByVal sender As Object, ByVal e As EventArgs) 
-        If My.Application.SaveMySettingsOnExit Then 
-            My.Settings.Save()
-        End If
-    End Sub
+    Private Shared addedHandlerLockObject As New Object
+
+    <Global.System.Diagnostics.DebuggerNonUserCodeAttribute(), Global.System.ComponentModel.EditorBrowsableAttribute(Global.System.ComponentModel.EditorBrowsableState.Advanced)> _
+    Private Shared Sub AutoSaveSettings(sender As Global.System.Object, e As Global.System.EventArgs)
+            If My.Application.SaveMySettingsOnExit Then
+                'My.Settings.Save()
+            End If
+        End Sub
 #End If
 #End Region
-    
-    Public Shared ReadOnly Property [Default]() As Settings
-        Get
-            
+        
+        Public Shared ReadOnly Property [Default]() As Settings
+            Get
+                
 #If _MyType = "WindowsForms" Then
-            If Not addedHandler Then
-                AddHandler My.Application.Shutdown, AddressOf AutoSaveSettings
-                addedHandler = True
-            End If
+               If Not addedHandler Then
+                    SyncLock addedHandlerLockObject
+                        If Not addedHandler Then
+                            AddHandler My.Application.Shutdown, AddressOf AutoSaveSettings
+                            addedHandler = True
+                        End If
+                    End SyncLock
+                End If
 #End If
-            Return defaultInstance
-        End Get
-    End Property
-End Class
+                Return defaultInstance
+            End Get
+        End Property
+    End Class
+End Namespace
 
 Namespace My
     
     <Global.Microsoft.VisualBasic.HideModuleNameAttribute(),  _
-     Global.System.Diagnostics.DebuggerNonUserCodeAttribute()>  _
-    Module MySettingsProperty
+     Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
+     Global.System.Runtime.CompilerServices.CompilerGeneratedAttribute()>  _
+    Friend Module MySettingsProperty
         
         <Global.System.ComponentModel.Design.HelpKeywordAttribute("My.Settings")>  _
-        Friend ReadOnly Property Settings() As Global.VbVbalrOOP.Settings
+        Friend ReadOnly Property Settings() As Global.VbVbalrOOP.My.Settings
             Get
-                Return Global.VbVbalrOOP.Settings.Default
+                Return Global.VbVbalrOOP.My.Settings.Default
             End Get
         End Property
     End Module
